Export functions individually instead
Rather than exporting everything as an object at the bottom of the
file, exporting the multiple functions individually appears to be a
better way for me to control which functions are made available outside
a module. This should come in handy when I eventually add helper
functions that are internal to a specific module.