Move `new` subcommand to program entry-point
From now on, ncg will have its subcommands declared under the main
program declaration rather than as standalone executables, and each
command will have an action handler that only calls a single function
which corresponds to the command's name (or feature). This is intended
to make the program modules easier to maintain, test and debug as the
declarations are grouped in one place while their underlying
implementation details are isolated.